move ntu course select schedule to data file
